backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
python-multipart==0.0.20
anthropic==0.43.0
Pillow==11.1.0
